lombok removal  related changes

+++ src/main/java/edu/wgu/dm/service/RoleService.java
@@ -11,19 +11,20 @@ import edu.wgu.dm.util.Permissions;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
-@RequiredArgsConstructor
 public class RoleService {
 
     private final RoleRepo roleRepo;
-
     private final UserRepo userRepo;
-
     private final PermissionRepo permRepo;
+
+    public RoleService(RoleRepo roleRepo, UserRepo userRepo, PermissionRepo permRepo) {
+        this.roleRepo = roleRepo;
+        this.userRepo = userRepo;
+        this.permRepo = permRepo;
+    }
 
     /**
      * Get All Role DTO
@@ -40,7 +41,10 @@ public class RoleService {
      * @param roleId
      * @return Role DTO
      */
-    public Role getRole(@NonNull Long roleId) {
+    public Role getRole(Long roleId) {
+        if (roleId == null) {
+            throw new NullPointerException("non null roleId is required");
+        }
         return this.roleRepo.getRoleById(roleId)
                             .orElseThrow(() -> new RoleNotFoundException(roleId));
     }
@@ -50,12 +54,16 @@ public class RoleService {
      *
      * @param roleId
      */
-    public void deleteRole(@NonNull Long roleId) {
+    public void deleteRole(Long roleId) {
+        if (roleId == null) {
+            throw new NullPointerException("non null roleId is required");
+        }
         this.roleRepo.deleteRole(roleId);
     }
 
     /**
-     * Add or update system Roles. Only allow users with the SYSTEM permission to create or update a role with the SYSTEM permission.
+     * Add or update system Roles. Only allow users with the SYSTEM permission to create or update a role with the
+     * SYSTEM permission.
      * <p>
      * Permissions need to exist before adding them to the role.
      *
@@ -63,7 +71,13 @@ public class RoleService {
      * @param roles
      * @return
      */
-    public List<Role> saveRoles(@NonNull String userId, @NonNull Role[] roles) {
+    public List<Role> saveRoles(String userId, Role[] roles) {
+        if (userId == null) {
+            throw new NullPointerException("non null userId is required");
+        }
+        if (roles == null) {
+            throw new NullPointerException("non null roles is required");
+        }
         List<Role> roleList = Arrays.asList(roles);
 
         List<Long> permissions = roleList.stream()
@@ -76,12 +90,19 @@ public class RoleService {
     }
 
     /**
-     * Validate --> If role has any system permission, only a system user can assign those. Find the SYSTEM permission. If any of the incoming permissions
-     * match, then check to see if the current user has the SYSTEM permission. If not, throw an AuthorizationException.
+     * Validate --> If role has any system permission, only a system user can assign those. Find the SYSTEM permission.
+     * If any of the incoming permissions match, then check to see if the current user has the SYSTEM permission. If
+     * not, throw an AuthorizationException.
      *
      * @param userId
      */
-    private void checkIfSystemUser(@NonNull List<Long> permissions, @NonNull String userId) {
+    private void checkIfSystemUser(List<Long> permissions, String userId) {
+        if (userId == null) {
+            throw new NullPointerException("non null userId is required");
+        }
+        if (permissions == null) {
+            throw new NullPointerException("non null permissions is required");
+        }
         Permission system = this.permRepo.getPermissionByName(Permissions.SYSTEM)
                                          .orElseThrow(() -> new IllegalStateException(
                                              "No SYSTEM permission is configured."));
